# website/static/openapi/v1.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vectis HTTP API",
    "version": "v1",
    "description": "Machine-readable contract for the shipped Vectis v1 HTTP API. Authentication can be disabled for local deployments; protected operations document bearer-token security because that is the production posture."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Default local vectis-api listener"
    }
  ],
...
  "paths": {
...
    "/api/v1/users/change-password": {
      "post": {
        "operationId": "changePassword",
        "summary": "Change the current user's password",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-vectis-auth-action": "api:any",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "operationId": "listUsers",
        "summary": "List local users",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-vectis-auth-action": "user:admin",
        "responses": {
          "200": {
            "description": "Local users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserResponse"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "createUser",
        "summary": "Create a local user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-vectis-auth-action": "user:admin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateUserResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "getUser",
        "summary": "Get a local user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-vectis-auth-action": "user:admin",
        "parameters": [
          {
            "$ref": "#/components/parameters/NumericID"
          }
        ],
        "responses": {
          "200": {
            "description": "Local user",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "put": {
        "operationId": "updateUser",
        "summary": "Update a local user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-vectis-auth-action": "user:admin",
        "parameters": [
          {
            "$ref": "#/components/parameters/NumericID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "delete": {
        "operationId": "deleteUser",
        "summary": "Delete a local user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-vectis-auth-action": "user:admin",
        "parameters": [
          {
            "$ref": "#/components/parameters/NumericID"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/users/{id}/external-identities": {
      "get": {
        "operationId": "listUserExternalIdentities",
        "summary": "List external identities linked to a user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-vectis-auth-action": "user:admin",
        "parameters": [
          {
            "$ref": "#/components/parameters/NumericID"
          }
        ],
        "responses": {
          "200": {
            "description": "External identity links",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ExternalIdentityResponse"
                  }
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "createUserExternalIdentity",
        "summary": "Link an external identity to a user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-vectis-auth-action": "user:admin",
        "parameters": [
          {
            "$ref": "#/components/parameters/NumericID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExternalIdentityRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "External identity linked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalIdentityResponse"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/v1/users/{id}/external-identities/{identity_id}": {
      "delete": {
        "operationId": "deleteUserExternalIdentity",
        "summary": "Unlink an external identity from a user",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-vectis-auth-action": "user:admin",
        "parameters": [
          {
            "$ref": "#/components/parameters/NumericID"
          },
          {
            "name": "identity_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/NoContent"
          },
          "default": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
      "NumericID": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer",
          "format": "int64",
          "minimum": 1
        }
      },
...
    },
...
    "responses": {
...
      "NoContent": {
        "description": "No response body"
      },
...
      "Error": {
        "description": "Vectis API error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/APIError"
            }
          }
        }
      }
    },
    "schemas": {
...
      "APIError": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ChangePasswordRequest": {
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "properties": {
          "current_password": {
            "type": "string",
            "format": "password"
          },
          "new_password": {
            "type": "string",
            "format": "password",
            "minLength": 8
          }
        },
        "additionalProperties": false
      },
      "CreateExternalIdentityRequest": {
        "type": "object",
        "required": [
          "provider_id",
          "subject"
        ],
        "properties": {
          "provider_id": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ExternalIdentityResponse": {
        "type": "object",
        "required": [
          "id",
          "provider_id",
          "subject",
          "local_user_id",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "provider_id": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "local_user_id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_login_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CreateUserRequest": {
        "type": "object",
        "required": [
          "username"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8
          }
        },
        "additionalProperties": false
      },
      "UpdateUserRequest": {
        "type": "object",
        "required": [
          "enabled"
        ],
        "properties": {
          "enabled": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "UserResponse": {
        "type": "object",
        "required": [
          "id",
          "username",
          "enabled",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "username": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "CreateUserResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserResponse"
          },
          {
            "type": "object",
            "properties": {
              "initial_password": {
                "type": "string"
              }
            }
          }
        ]
      },
...
    }
  }
}